sp-io and externalities cleanup
