+++ substrate-sgx/externalities/Cargo.toml
@@ -5,13 +5,11 @@ authors = ["Integritee AG <hello@integritee.network> and Parity Technologies <ad
 edition = "2018"
 
 [dependencies]
-
 # no_std
 codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive", "chain-error"]}
 derive_more = "0.99.16"
 environmental = { version = "1.1.3", default-features = false }
 log = { version = "0.4", default-features = false }
-postcard = { version = "0.7", features = ["alloc"] }
 serde = { version = "1.0", default-features = false, features = ["derive", "alloc"] }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 
@@ -25,7 +23,6 @@ std = [
     "codec/std",
     "environmental/std",
     "log/std",
-    "postcard/use-std",
     "serde/std",
     "serde_json/std",
 ]
