+++ substrate-sgx/sp-io/Cargo.toml
@@ -7,30 +7,14 @@ license = "Apache-2.0"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
-hash-db = { version = "0.15.2", default-features = false }
-libsecp256k1 = { version = "0.6.0", optional = true }
-futures = { version = "0.3.1", features = ["thread-pool"], optional = true }
-parking_lot = { version = "0.11.1", optional = true }
-tracing = { version = "0.1.25", default-features = false }
-tracing-core = { version = "0.1.17", default-features = false}
 log = { version = "0.4", default-features = false }
 
-environmental = { version = "1.1.3", default-features = false }
 sgx_tstd      = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["untrusted_fs","net","backtrace"], optional = true}
 sgx_types     = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true}
-#sgx_log       = { package = "log", git = "https://github.com/mesalock-linux/log-sgx", optional =true }
 sgx-externalities = { default-features = false, path = "../externalities", optional = true }
 
 # Substrate dependencies
-sp-std = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 sp-core = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", features=["full_crypto"], branch = "master" }
-sp-state-machine = { version = "0.10.0-dev", git = "https://github.com/paritytech/substrate.git", branch = "master", optional = true}
-sp-runtime-interface = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
-sp-wasm-interface = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
-sp-tracing = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
-sp-trie = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate.git", branch = "master", optional = true }
-sp-keystore = {  version = "0.10.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master", optional = true }
-sp-externalities = { version = "0.10.0-dev", git = "https://github.com/paritytech/substrate.git", branch = "master", optional = true }
 
 [dev-dependencies]
 hex-literal = { version = "0.3.4" }
@@ -39,21 +23,9 @@ hex-literal = { version = "0.3.4" }
 default = ["std"]
 
 std = [
-	"environmental/std",
 	"log/std",
 	"sp-core/std",
 	"codec/std",
-	"sp-std/std",
-	"sp-keystore/std",
-	"hash-db/std",
-	"sp-trie",
-	"sp-state-machine",
-	"libsecp256k1",
-	"sp-runtime-interface/std",
-	"sp-externalities",
-	"sp-wasm-interface/std",
-	"futures",
-	"parking_lot",
 	"sgx-externalities/std",
 ]
 
@@ -61,13 +33,4 @@ sgx = [
 	"sgx_tstd",
 	"sgx_types",
 	"sgx-externalities/sgx",
-	"sp-runtime-interface/disable_target_static_assertions",
 ]
-
-# These two features are used for `no_std` builds for the environments which already provides
-# `#[panic_handler]`, `#[alloc_error_handler]` and `#[global_allocator]`.
-#
-# For the regular wasm runtime builds those are not used.
-disable_panic_handler = []
-disable_oom = []
-disable_allocator = []
