+++ test_no_std/Cargo.toml
@@ -23,7 +23,7 @@ default-features = false
 [dependencies.sp-io]
 path = "../substrate-sgx/sp-io"
 default-features = false
-features = ["disable_oom", "disable_panic_handler", "disable_allocator", "sgx"]
+features = ["sgx"]
 
 [dependencies.sp-application-crypto]
 git = "https://github.com/paritytech/substrate.git"
